+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - **Added:** Disable SyntaxWarnings to reduce output noise from python parser.
 - **Added:** `DOC008` One-line docstring should end with a period.
-- **Fixed:** `DOC008` Ignore multi-line docstrings with ignored sections or invalid sections.
+- **Added:** `DOC009` Docstring should use """triple double quotes"""
 
 ### v0.0.6 (2025-09-25)